+++ chakracore-cxx/lib/Common/BackendApi.h
@@ -193,32 +193,3 @@ struct ThunkConstants
     static const size_t ThunkSize = 6;
 };
 #endif
-
-enum LibraryValue {
-    ValueInvalid,
-    ValueUndeclBlockVar,
-    ValueEmptyString,
-    ValueUndefined,
-    ValueNull,
-    ValueTrue,
-    ValueFalse,
-    ValueNegativeZero,
-    ValueNumberTypeStatic,
-    ValueStringTypeStatic,
-    ValueSymbolTypeStatic,
-    ValueObjectType,
-    ValueObjectHeaderInlinedType,
-    ValueRegexType,
-    ValueArrayConstructor,
-    ValuePositiveInfinity,
-    ValueNaN,
-    ValueJavascriptArrayType,
-    ValueNativeIntArrayType,
-    ValueNativeFloatArrayType,
-    ValueConstructorCacheDefaultInstance,
-    ValueAbsDoubleCst,
-    ValueUintConvertConst,
-    ValueBuiltinFunctions,
-    ValueJnHelperMethods,
-    ValueCharStringCache
-};
