+++ chakracore-cxx/lib/Runtime/Library/JavascriptArray.h
@@ -5,6 +5,7 @@
 //-------------------------------------------------------------------------------------------------------
 
 #pragma once
+#include "Common/LibraryValue.h"
 
 #define ARRAY_CROSSOVER_FOR_VALIDATE 0
